feat: add useValidateCardInfos hook
